[CHANGE] implemented recent document context menu on Document HUBs recent documents disclosure item.

+++ SubEthaEdit-Mac/NSWorkspaceTCMAdditions.h
@@ -10,5 +10,5 @@
 
 
 @interface NSWorkspace (NSWorkspaceTCMAdditions) 
-- (NSImage *)iconForFileType:(NSString *)anExtension size:(int)aSize;
+- (NSImage *)iconForFileType:(NSString *)anExtension size:(NSInteger)aSize;
 @end
